big integer

+++ Assets/Clicker/Scripts/Domain/Model/Wallet.cs
@@ -1,24 +1,25 @@
 using System.Collections;
 using System.Collections.Generic;
+using System.Numerics;
 
 namespace Clicker
 {
     public class Wallet
     {
-        public int Coin { get; private set; }
+        public BigInteger Coin { get; private set; }
 
-        public Wallet(int coin) {
+        public Wallet(BigInteger coin) {
             this.Coin = coin;
         }
 
-        public void ConsumCoin(int coin) {
+        public void ConsumCoin(BigInteger coin) {
             if (this.Coin < coin) {
                 throw new System.Exception("コインが足りません");//TODO
             }
             this.Coin -= coin;
         }
 
-        public void AddCoin(int coin) {
+        public void AddCoin(BigInteger coin) {
             this.Coin += coin;
         }
     }
